requirements.txt
-f /usr/share/pip-wheels
akismet==1.1
asgiref==3.4.1
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
crispy-bootstrap5==0.5
cssselect2==0.4.1
Django==3.2.6
django-address==0.2.5
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-bower==5.2.0
django-compressor==2.4.1
django-contact-form==1.9
django-crispy-forms==1.12.0
django-environ==0.8.1
django-excel-response==2.0.5
django-extensions==3.1.3
django-guardian==2.4.0
django-imperavi-widget==0.0.3
django-jquery==3.1.0
django-jquery-ui==1.11.4.1
django-multipleselectfield==0.2.0.post2
django-newsletter==0.9.1
django-phone-field==1.8.1
django-photologue==3.14
django-ranged-response==0.2.0
django-renderpdf==3.0.1
django-sass-processor==1.0.1
django-scheduler==0.9.5
django-simple-captcha==0.5.14
django-sortedm2m==3.1.1
django-tinymce==3.3.0
django-userena-ce==6.0.1
easy-thumbnails==2.7.1
et-xmlfile==1.1.0
ExifRead==2.3.2
fonttools==4.27.1
html2text==2020.1.16
html5lib==1.1
icalendar==4.0.7
idna==3.2
jdcal==1.4.1
ldif3==3.1.1
libsass==0.21.0
Markdown==3.3.4
npm==0.1.1
openpyxl==2.4.11
optional-django==0.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
pydyf==0.1.1
pyphen==0.11.0
python-card-me==0.9.3
python-dateutil==2.8.2
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
surlex==0.2.0
tinycss2==1.1.0
unicodecsv==0.14.1
urllib3==1.26.6
weasyprint==53.3
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.1.8
